Added a conditional check to TranslateData to prevent unnecesary copying
The hashtable returned from TranslateData_Impl could be the provided hashtable.
Because of this it is a good idea to check for reference equality so that we
don't copy a reference to an image into the same cell. While the transfer may
be simple, I don't know if there is anything strange that occurs. Thus, it is a
good idea to just do the equality check first to see if it's even necessary