Fixed a bug in the standard save function
I found that the check to make sure that the resultant image was not null would
never work correctly because it was checking to see if the image display panel
was null, not the image it contained.

I have fixed this and now the correct message is displayed if you haven't set a
resultant image to be saved.