Migrated filters to support color
Filter remains as the class for grey scale only images though. Instead of using
byte[][], the full color versions use int[][].

Filter now converts to and from this new format automatically so that
pre-existing plugins are unaware of the change. It does mean that greyscale
plugins take a hit in performance though.