Updated Filter's TransformImage method to support color filters
Since filters originally operated on pixel intensity, we can apply the same
filter three times (once for each color channel) and then merge the image back
together. This is super inefficient but this is done for compatibility reasons.

It is up to the plugin author to rewrite the filter to take advantage of
colors.